requirements.txt
# Python dependencies
python-dotenv==1.0.0
Flask==3.0.0
